refactors separate api requests from location search.
