+++ app/controllers/api/v1/background_controller.rb
@@ -1,6 +1,6 @@
 class Api::V1::BackgroundController < ApplicationController
   def index
-    result = LocationSearch.new(params[:location]).get_photos
+    result = FlickrRequest.new(params[:location]).get_photos
     render json: result
   end
 end
