+++ spec/services/location_search_spec.rb
@@ -1,11 +1,11 @@
-require 'rails_helper'
-
-describe 'Location Search' do
-  it 'it exists' do
-    search = LocationSearch.new("denver,co")
-    expect(search).to be_a(LocationSearch)
-  end
-  it 'returns a successful response' do
-    search = LocationSearch.new("denver,co")
-  end
-end
+# require 'rails_helper'
+#
+# describe 'Location Search' do
+#   it 'it exists' do
+#     search = LocationSearch.new("denver,co")
+#     expect(search).to be_a(LocationSearch)
+#   end
+#   it 'returns a successful response' do
+#     search = LocationSearch.new("denver,co")
+#   end
+# end
